Renaming IMetadataConfigurationApi to IMetadataApiConfiguration
Renaming IPOEditorApi to IPOApiEditor

Moving IMetadataConfigurationDataSource.java

Renaming Injector to AppFactory
