+++ app/src/androidTestereferrals/java/org/eyeseetea/malariacare/MetadataConfigurationDBImporterShould.java
@@ -16,7 +16,7 @@ import org.eyeseetea.malariacare.configurationImporter.BaseMetadataConfiguration
 import org.eyeseetea.malariacare.data.database.model.OptionDB;
 import org.eyeseetea.malariacare.data.database.model.QuestionDB;
 import org.eyeseetea.malariacare.data.database.model.QuestionOptionDB;
-import org.eyeseetea.malariacare.data.di.Injector;
+import org.eyeseetea.malariacare.data.di.AppFactory;
 import org.eyeseetea.malariacare.data.sync.importer.metadata.configuration.MetadataConfigurationDBImporter;
 import org.junit.Assert;
 import org.junit.Before;
@@ -45,7 +45,7 @@ public class MetadataConfigurationDBImporterShould extends BaseMetadataConfigura
         shouldNotBeAnyQuestionInTheDB();
 
         MetadataConfigurationDBImporter importer = new MetadataConfigurationDBImporter(
-                Injector.provideQuestionConverter()
+                AppFactory.provideQuestionConverter()
         );
 
         try {
